requirements.txt
torch
torchvision
torchaudio
streamlit
pillow
numpy
pandas
scikit-learn
tqdm
pydantic
